[Console] add setInputs to ApplicationTest and share some code
