+++ src/Symfony/Component/Console/Tester/CommandTester.php
@@ -14,8 +14,6 @@ namespace Symfony\Component\Console\Tester;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Output\StreamOutput;
-use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Output\OutputInterface;
 
 /**
  * Eases the testing of console commands.
@@ -25,10 +23,10 @@ use Symfony\Component\Console\Output\OutputInterface;
  */
 class CommandTester
 {
+    use TesterTrait;
+
     private $command;
     private $input;
-    private $output;
-    private $inputs = array();
     private $statusCode;
 
     public function __construct(Command $command)
@@ -77,80 +75,5 @@ class CommandTester
         }
 
         return $this->statusCode = $this->command->run($this->input, $this->output);
-    }
-
-    /**
-     * Gets the display returned by the last execution of the command.
-     *
-     * @param bool $normalize Whether to normalize end of lines to \n or not
-     *
-     * @return string The display
-     */
-    public function getDisplay($normalize = false)
-    {
-        rewind($this->output->getStream());
-
-        $display = stream_get_contents($this->output->getStream());
-
-        if ($normalize) {
-            $display = str_replace(PHP_EOL, "\n", $display);
-        }
-
-        return $display;
-    }
-
-    /**
-     * Gets the input instance used by the last execution of the command.
-     *
-     * @return InputInterface The current input instance
-     */
-    public function getInput()
-    {
-        return $this->input;
-    }
-
-    /**
-     * Gets the output instance used by the last execution of the command.
-     *
-     * @return OutputInterface The current output instance
-     */
-    public function getOutput()
-    {
-        return $this->output;
-    }
-
-    /**
-     * Gets the status code returned by the last execution of the application.
-     *
-     * @return int The status code
-     */
-    public function getStatusCode()
-    {
-        return $this->statusCode;
-    }
-
-    /**
-     * Sets the user inputs.
-     *
-     * @param array $inputs An array of strings representing each input
-     *                      passed to the command input stream
-     *
-     * @return CommandTester
-     */
-    public function setInputs(array $inputs)
-    {
-        $this->inputs = $inputs;
-
-        return $this;
-    }
-
-    private static function createStream(array $inputs)
-    {
-        $stream = fopen('php://memory', 'r+', false);
-
-        fwrite($stream, implode(PHP_EOL, $inputs));
-        rewind($stream);
-
-        return $stream;
     }
 }
