+++ src/Symfony/Component/Console/Tests/Tester/ApplicationTesterTest.php
@@ -13,7 +13,9 @@ namespace Symfony\Component\Console\Tests\Tester;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Application;
+use Symfony\Component\Console\Helper\QuestionHelper;
 use Symfony\Component\Console\Output\Output;
+use Symfony\Component\Console\Question\Question;
 use Symfony\Component\Console\Tester\ApplicationTester;
 
 class ApplicationTesterTest extends TestCase
@@ -27,7 +29,9 @@ class ApplicationTesterTest extends TestCase
         $this->application->setAutoExit(false);
         $this->application->register('foo')
             ->addArgument('foo')
-            ->setCode(function ($input, $output) { $output->writeln('foo'); })
+            ->setCode(function ($input, $output) {
+                $output->writeln('foo');
+            })
         ;
 
         $this->tester = new ApplicationTester($this->application);
@@ -61,6 +65,25 @@ class ApplicationTesterTest extends TestCase
     public function testGetDisplay()
     {
         $this->assertEquals('foo'.PHP_EOL, $this->tester->getDisplay(), '->getDisplay() returns the display of the last execution');
+    }
+
+    public function testSetInputs()
+    {
+        $application = new Application();
+        $application->setAutoExit(false);
+        $application->register('foo')->setCode(function ($input, $output) {
+            $helper = new QuestionHelper();
+            $helper->ask($input, $output, new Question('Q1'));
+            $helper->ask($input, $output, new Question('Q2'));
+            $helper->ask($input, $output, new Question('Q3'));
+        });
+        $tester = new ApplicationTester($application);
+
+        $tester->setInputs(array('I1', 'I2', 'I3'));
+        $tester->run(array('command' => 'foo'));
+
+        $this->assertSame(0, $tester->getStatusCode());
+        $this->assertEquals('Q1Q2Q3', $tester->getDisplay(true));
     }
 
     public function testGetStatusCode()
